+++ src/vue/forms/PayForm.vue
@@ -69,6 +69,7 @@ export default {
         amount: '',
         quoteAssetCode: '',
         paymentMethodId: '',
+        promoCode: '',
       },
       isDisabled: false,
       atomicSwapBidDetails: {
@@ -117,6 +118,7 @@ export default {
           this.form.amount,
           this.form.paymentMethodId,
           this.atomicSwapAsk.id,
+          this.form.promoCode,
           this.form.email
         )
         if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {