+++ src/vue/modules/atomic-swap-form/index.vue
@@ -62,6 +62,7 @@ export default {
         amount: '',
         quoteAssetCode: '',
         paymentMethodId: '',
+        promoCode: '',
       },
       isDisabled: false,
       atomicSwapBidDetails: {
@@ -91,7 +92,8 @@ export default {
         const atomicSwapBid = await this.createAtomicSwapBidOperation(
           this.form.amount,
           this.form.paymentMethodId,
-          this.atomicSwapAsk.id
+          this.atomicSwapAsk.id,
+          this.form.promoCode
         )
         if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {
           window.location.href = atomicSwapBid.payUrl